Practice planning questions answered.
I probably need to get more specific in some areas, and I still need
to write more user stories.
When you say "write 3-5", are those drafts?  Or just 3-5 user stories
based on what users should be able to do in the app?
Also, how much time should it realistically take for me to get the
back-end endpoints up and running?  Is my timetable realistic? I left
it unfinished on purpose, because I'm not entirely sure if it is
realistic.